// src/renderer/tools/types/IMicroInputType.ts
export enum IMicroInputType {
  VALUE = 'value',
	NAME = 'name',
	START_BRACKET = 'startBracket',
	END_BRACKET = 'endBracket',
	START_ADD_BRACKET = 'startAddBracket',
	END_ADD_BRACKET = 'endAddBracket',
  OPERATION = 'operation',
  SEPARATOR = 'separator',
  DEGREE = 'degree',
	ANSWER = 'answer',
}